webapi/requirements.txt
Flask==0.11
Flask-Bootstrap==3.3.6.0
Flask-SocketIO==2.4
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.10
argparse==1.2.1
click==6.6
dominate==2.2.0
eventlet==0.19.0
greenlet==0.4.9
itsdangerous==0.24
kafka-python==1.2.1
psycopg2==2.6.1
python-engineio==0.9.1
python-socketio==1.3
six==1.10.0
visitor==0.1.3
wsgiref==0.1.2
